requirements.txt
uwsgi
markerlib
defusedxml
flask
flask-wtf
nltk
python-openid
requests
textblob
psutil
python-dotenv